handeled stuff mentioned
